Extend provision task to also create the `players` table
The trick to avoiding a pyramid of doom was to leverage `Promise.bind`
to pass the open connection through sequential `then` promises (bind is
necessary because rethink's `run` returns the results).

A nice side effect is that ES6's arrow function retains the `this`
binding, so the nested rethink commands inside the create function
retain access to `this.connection`.

http://bluebirdjs.com/docs/api/promise.bind.html