Implement update player endpoint against db
I also realized that the controllers should respond during the `then`
handler of the command, instead of `finally`, so that they don't
respond with the player object if something unexpected happens.

I also changed the error code for updating or creating a player with an
incomplete model from 404 Not Found to 400 Bad Request.