Move middleware out of `index.js`
Since I'll need to add more middleware in the future (at the least, I
want a close-connection piece) I created a new place in the project to
put it.

Also, since I wanted the middleware folder to be a sibling of the
folder with the endpoints in it *and* the middleware is "part of" the
api in my mind, I renamed that folder from `api` to `endpoints`.