Add phonetic name field and split apart change and update services
I wanted to test for the case where a consumer POSTs to a player with
an object containing id, but changed the id for some reason. Because
this is a potential gotcha for `update`, but not for `create` (because
consumers don't specify an id for creates), I needed to separate the
logic.