Update provision task to create api_keys table
This was tricky; my first thought was to change `createTable` to accept
a parameter of the table name and chain a few `.then`s invoking this
new function.

However, a construct like `.then(createTable('players')` was kicking
off the create promise immediately, so the chain became asynchronous
and the connection wasn't getting passed between handlers correctly.

The solution was to use partial application to create functions
specific to each parameter. This way the promises go back to executing
one at a time because they aren't *invoked* immediately.