Implement delete singles match endpoint
I also noticed the delete.map service did not have unit tests, so I
added them for the player service as well.

It is starting to look like some of these services can be combined; for
example, players and matches both have a generic `find` validator that
just checks to make sure the provided `id` is a valid guid.